Move test - changed init and moved config for grabbing info
