+++ app/dbConfig.py
@@ -1,0 +1,14 @@
+# Replace with the correct username
+ssh_username = 'ubuntu'
+
+
+#Replace with where your .pem file location is
+key_Path = ''
+
+
+USER='admin'
+#Repalce with the password
+PASS=''
+HOST='127.0.0.1'
+#Change to the db name
+MYDB=''
